refactor(adapter-web,utils): consolidate OPFS debug utilities
- Remove deprecated `opfs-utils.ts` from adapter-web package
- Move and enhance debug utilities into `@livestore/utils/effect/Opfs`
- Add new `debug-utils.ts` with improved tree visualization including file metadata
- Update all references to use `Opfs.debugUtils` instead of old OpfsUtils
- Remove opfs-utils export from adapter-web package.json
