+++ packages/@livestore/utils/src/effect/Opfs/index.ts
@@ -1,2 +1,3 @@
+export * from './debug-utils.ts'
 export * from './Opfs.ts'
 export * from './utils.ts'
