+++ packages/@livestore/utils/src/effect/Opfs/utils.ts
@@ -273,123 +273,3 @@ export const syncWriteFile = Effect.fn('@livestore/utils:Opfs.syncWriteFile')(fu
 
   yield* Opfs.syncFlush(handle)
 })
-
-/**
- * Metadata exposed directly on OPFS handles so we avoid `getFile()` reads.
- */
-interface OpfsEntryMetadata {
-  readonly name: string
-  readonly path: string
-  readonly kind: FileSystemHandleKind
-}
-
-interface OpfsTreeNode {
-  readonly metadata: OpfsEntryMetadata
-  readonly children?: ReadonlyArray<OpfsTreeNode>
-}
-
-const ROOT_NAME = '/'
-
-/**
- * Materialize the entire OPFS tree starting from the origin root.
- */
-const buildTree = Effect.fn('@livestore/utils:Opfs.buildTree')(function* () {
-  const rootHandle = yield* Opfs.getRootDirectoryHandle
-
-  const collectDirectory: (
-    handle: FileSystemDirectoryHandle,
-    pathSegments: ReadonlyArray<string>,
-  ) => Effect.Effect<OpfsTreeNode, unknown, Opfs> = (handle, pathSegments) =>
-    Effect.gen(function* () {
-      const entries = yield* Opfs.listEntries(handle)
-      const sorted = entries.slice().sort((left, right) => left.name.localeCompare(right.name))
-
-      const children: OpfsTreeNode[] = []
-
-      for (const entry of sorted) {
-        const nextSegments = [...pathSegments, entry.name]
-        const path = formatPath(nextSegments)
-
-        if (entry.kind === 'directory') {
-          const childDirectory = yield* collectDirectory(entry.handle, nextSegments)
-          children.push(childDirectory)
-          continue
-        }
-
-        children.push({
-          metadata: {
-            name: entry.name,
-            path,
-            kind: entry.kind,
-          },
-        })
-      }
-
-      return {
-        metadata: {
-          name: pathSegments.length === 0 ? ROOT_NAME : pathSegments[pathSegments.length - 1]!,
-          path: formatPath(pathSegments),
-          kind: 'directory',
-        },
-        children,
-      }
-    })
-
-  return yield* collectDirectory(rootHandle, [])
-})
-
-const formatPath = (segments: ReadonlyArray<string>) => (segments.length === 0 ? ROOT_NAME : `/${segments.join('/')}`)
-
-const formatLabel = ({ name, kind }: OpfsEntryMetadata) => `${name}${kind === 'directory' ? ' [dir]' : ` [${kind}]`}`
-
-const logAsciiTree = (node: OpfsTreeNode): Effect.Effect<void, never, never> =>
-  logAsciiNode(node, { prefix: '', isLast: true, isRoot: true })
-
-const logAsciiNode: (
-  node: OpfsTreeNode,
-  options: { readonly prefix: string; readonly isLast: boolean; readonly isRoot?: boolean },
-) => Effect.Effect<void, never, never> = (node, options) =>
-  Effect.gen(function* () {
-    const label = formatLabel(node.metadata)
-    const branch = options.isRoot ? '' : `${options.prefix}${options.isLast ? '└── ' : '├── '}`
-    const nextPrefix = options.isRoot ? '' : `${options.prefix}${options.isLast ? '    ' : '│   '}`
-
-    yield* Effect.log(`${branch}${label}`)
-
-    if (node.children === undefined || node.children.length === 0) return
-
-    for (let index = 0; index < node.children.length; index++) {
-      const child = node.children[index]!
-      const isLastChild = index === node.children.length - 1
-      yield* logAsciiNode(child, { prefix: nextPrefix, isLast: isLastChild })
-    }
-  })
-
-const logFlatTree: (node: OpfsTreeNode) => Effect.Effect<void, never, never> = (node) =>
-  Effect.gen(function* () {
-    yield* Effect.log(`${node.metadata.path} [${node.metadata.kind}]`)
-
-    if (node.children === undefined || node.children.length === 0) return
-
-    for (const child of node.children) {
-      yield* logFlatTree(child)
-    }
-  })
-
-const logJsonTree: (node: OpfsTreeNode) => Effect.Effect<void, never, never> = (node) =>
-  Effect.log(JSON.stringify(node, null, 2))
-
-export const logOpfsTreeAscii = Effect.fn('@livestore/utils:Opfs.logOpfsTreeAscii')(function* () {
-  const tree = yield* buildTree()
-  yield* logAsciiTree(tree)
-})
-
-export const logOpfsTreeFlatPaths = Effect.fn('@livestore/utils:Opfs.logOpfsTreeFlatPaths')(function* () {
-  const tree = yield* buildTree()
-  yield* logFlatTree(tree)
-})
-
-export const logOpfsTreeJson = Effect.fn('@livestore/utils:Opfs.logOpfsTreeJson')(function* () {
-  const tree = yield* buildTree()
-  yield* logJsonTree(tree)
-})
